// host-manager/src/interfaces/farm.ts
export interface VectorKeyObjectValue {
    key: KeyByVector;
    value: ObjectValue;
}


export interface ObjectValue {
    Object: Object;
}
  
export interface Object {
    "@_xsi:type"?: string;
    category: number;
    name: string;
    parentSheetIndex: number;
    specialItem: boolean;
    hasBeenInInventory: boolean;
    SpecialVariable: number;
    Category: number;
    ParentSheetIndex: number;
    DisplayName: string;
    Name: string;
    Stack: number;
    tileLocation: Tile;
    owner: number;
    type: string;
    canBeSetDown: boolean;
    canBeGrabbed: boolean;
    isHoedirt: boolean;
    isSpawnedObject: boolean;
    questItem: boolean;
    questId: number;
    isOn: boolean;
    fragility: number;
    price: number;
    edibility: number;
    stack: number;
    quality: number;
    bigCraftable: boolean;
    setOutdoors: boolean;
    setIndoors: boolean;
    readyForHarvest: boolean;
    showNextIndex: boolean;
    flipped: boolean;
    hasBeenPickedUpByFarmer: boolean;
    isRecipe: boolean;
    isLamp: boolean;
    minutesUntilReady: number;
    boundingBox: BoundingBox;
    scale: Tile;
    preservedParentSheetIndex: number;
    health?: number;
    maxHealth?: number;
    whichType?: number;
    gatePosition?: number;
    gateMotion?: number;
    isGate?: boolean;
    heldObject?: HeldObject;
    honeyType?: string;
    currentLidFrame?: number;
    frameCounter?: number;
    coins?: number;
    items?: ObjectItems;
    chestType?: string;
    tint?: Tint;
    playerChoiceColor?: Tint;
    playerChest?: boolean;
    fridge?: boolean;
    giftbox?: boolean;
    Tint?: Tint;
  }

 

  export interface ObjectItems {
    Item: any;
  }


  export interface LargeTerrainFeatures {
    LargeTerrainFeature: LargeTerrainFeature[];
  }
  
export interface LargeTerrainFeature {
    "@_xsi:type": string;
    tilePosition: Tile;
    size: number;
    tileSheetOffset: number;
    health: number;
    flipped: boolean;
    townBush: boolean;
    drawShadow: boolean;
}

export interface TerrainFeatures {
    item: VectorKeyTerrainFeatureValue[];
}
  
export interface VectorKeyTerrainFeatureValue {
    key: KeyByVector;
    value: TerrainFeatureValue;
}

export interface TerrainFeatureValue {
    TerrainFeature: TerrainFeature;
}
  
export interface TerrainFeature {
    "@_xsi:type": string;
    growthStage?: number;
    treeType?: number;
    health?: number;
    flipped?: boolean;
    stump?: boolean;
    tapped?: boolean;
    hasSeed?: boolean;
    shakeLeft?: boolean;
    grassType?: number;
    numberOfWeeds?: number;
    grassSourceOffset?: number;
    whichFloor?: number;
    whichView?: number;
    indexOfFruit?: number;
    daysUntilMature?: number;
    fruitsOnTree?: number;
    struckByLightningCountdown?: number;
    greenHouseTree?: boolean;
    greenHouseTileTree?: boolean;
    fruitSeason?: string;
    state?: number;
    fertilizer?: number;
    crop?: Crop;
}
  
export interface Crop {
    phaseDays: PhaseDays;
    rowInSpriteSheet: number;
    phaseToShow: number;
    currentPhase: number;
    harvestMethod: number;
    indexOfHarvest: number;
    regrowAfterHarvest: number;
    dayOfCurrentPhase: number;
    minHarvest: number;
    maxHarvest: number;
    maxHarvestIncreasePerFarmingLevel: number;
    daysOfUnclutteredGrowth: number;
    whichForageCrop: number;
    seasonsToGrowIn: SeasonsToGrowIn;
    tintColor: Tint;
    flip: boolean;
    fullGrown: boolean;
    raisedSeeds: boolean;
    programColored: boolean;
    dead: boolean;
    forageCrop: boolean;
    chanceForExtraCrops: number;
}
  
export interface PhaseDays {
    int: number[];
}
  
export interface SeasonsToGrowIn {
    string: string;
}